requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
azure-cosmos>=4.5.1
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
pytz>=2023.3
requests>=2.31.0
plotly>=5.17.0
numpy>=1.24.0
pydantic>=2.5.0
aiohttp>=3.8.0
typing-extensions>=4.8.0
gunicorn>=21.0.0
